backend/requirements.txt
fastapi[standard]
uvicorn

psycopg2-binary
python-multipart

pandas
openpyxl

python-docx
docxtpl